+++ sdk/metricsadvisor/Azure.AI.MetricsAdvisor/tests/MetricsAdvisorAdministrationClient/DataFeedLiveTests.cs
@@ -3,6 +3,7 @@
 
 using System;
 using System.Linq;
+using System.Reflection;
 using System.Threading.Tasks;
 using Azure.AI.MetricsAdvisor.Administration;
 using Azure.AI.MetricsAdvisor.Models;
@@ -487,11 +488,13 @@ namespace Azure.AI.MetricsAdvisor.Tests
         }
 
         [RecordedTest]
-        [TestCase(AzureBlobDataFeedSource.AuthenticationType.Basic)]
-        [TestCase(AzureBlobDataFeedSource.AuthenticationType.ManagedIdentity)]
-        public async Task CreateAndGetAzureBlobDataFeedWithAuthentication(AzureBlobDataFeedSource.AuthenticationType authentication)
+        [TestCase("Basic")]
+        [TestCase("ManagedIdentity")]
+        public async Task CreateAndGetAzureBlobDataFeedWithAuthentication(string authenticationType)
         {
             MetricsAdvisorAdministrationClient adminClient = GetMetricsAdvisorAdministrationClient();
+            AzureBlobDataFeedSource.AuthenticationType authentication =
+                GetAuthenticationInstance<AzureBlobDataFeedSource.AuthenticationType>(authenticationType);
 
             var dataFeedName = Recording.GenerateAlphaNumericId("dataFeed");
             var dataSource = new AzureBlobDataFeedSource("mock", "mock", "mock")
@@ -509,11 +512,13 @@ namespace Azure.AI.MetricsAdvisor.Tests
         }
 
         [RecordedTest]
-        [TestCase(AzureDataExplorerDataFeedSource.AuthenticationType.Basic)]
-        [TestCase(AzureDataExplorerDataFeedSource.AuthenticationType.ManagedIdentity)]
-        public async Task CreateAndGetAzureDataExplorerDataFeedWithAuthentication(AzureDataExplorerDataFeedSource.AuthenticationType authentication)
+        [TestCase("Basic")]
+        [TestCase("ManagedIdentity")]
+        public async Task CreateAndGetAzureDataExplorerDataFeedWithAuthentication(string authenticationType)
         {
             MetricsAdvisorAdministrationClient adminClient = GetMetricsAdvisorAdministrationClient();
+            AzureDataExplorerDataFeedSource.AuthenticationType authentication =
+                GetAuthenticationInstance<AzureDataExplorerDataFeedSource.AuthenticationType>(authenticationType);
 
             var dataFeedName = Recording.GenerateAlphaNumericId("dataFeed");
             var dataSource = new AzureDataExplorerDataFeedSource("mock", "mock")
@@ -531,14 +536,16 @@ namespace Azure.AI.MetricsAdvisor.Tests
         }
 
         [RecordedTest]
-        [TestCase(AzureDataExplorerDataFeedSource.AuthenticationType.ServicePrincipal)]
-        [TestCase(AzureDataExplorerDataFeedSource.AuthenticationType.ServicePrincipalInKeyVault)]
-        public async Task CreateAndGetAzureDataExplorerDataFeedWithCredentialAuthentication(AzureDataExplorerDataFeedSource.AuthenticationType authentication)
+        [TestCase("ServicePrincipal")]
+        [TestCase("ServicePrincipalInKeyVault")]
+        public async Task CreateAndGetAzureDataExplorerDataFeedWithCredentialAuthentication(string authenticationType)
         {
             MetricsAdvisorAdministrationClient adminClient = GetMetricsAdvisorAdministrationClient();
+            AzureDataExplorerDataFeedSource.AuthenticationType authentication =
+                GetAuthenticationInstance<AzureDataExplorerDataFeedSource.AuthenticationType>(authenticationType);
 
             var credentialName = Recording.GenerateAlphaNumericId("credential");
-            await using var disposableCredential = await DisposableDataSourceCredentialEntity.CreateDataSourceCredentialEntityAsync(adminClient, credentialName, authentication.ToString());
+            await using var disposableCredential = await DisposableDataSourceCredentialEntity.CreateDataSourceCredentialEntityAsync(adminClient, credentialName, authenticationType);
             string credentialId = disposableCredential.Credential.Id;
 
             var dataFeedName = Recording.GenerateAlphaNumericId("dataFeed");
@@ -579,15 +586,17 @@ namespace Azure.AI.MetricsAdvisor.Tests
         }
 
         [RecordedTest]
-        [TestCase(AzureDataLakeStorageDataFeedSource.AuthenticationType.ServicePrincipal)]
-        [TestCase(AzureDataLakeStorageDataFeedSource.AuthenticationType.ServicePrincipalInKeyVault)]
-        [TestCase(AzureDataLakeStorageDataFeedSource.AuthenticationType.SharedKey)]
-        public async Task CreateAndGetAzureDataLakeStorageDataFeedWithCredentialAuthentication(AzureDataLakeStorageDataFeedSource.AuthenticationType authentication)
+        [TestCase("ServicePrincipal")]
+        [TestCase("ServicePrincipalInKeyVault")]
+        [TestCase("SharedKey")]
+        public async Task CreateAndGetAzureDataLakeStorageDataFeedWithCredentialAuthentication(string authenticationType)
         {
             MetricsAdvisorAdministrationClient adminClient = GetMetricsAdvisorAdministrationClient();
+            AzureDataLakeStorageDataFeedSource.AuthenticationType authentication =
+                GetAuthenticationInstance<AzureDataLakeStorageDataFeedSource.AuthenticationType>(authenticationType);
 
             var credentialName = Recording.GenerateAlphaNumericId("credential");
-            await using var disposableCredential = await DisposableDataSourceCredentialEntity.CreateDataSourceCredentialEntityAsync(adminClient, credentialName, authentication.ToString());
+            await using var disposableCredential = await DisposableDataSourceCredentialEntity.CreateDataSourceCredentialEntityAsync(adminClient, credentialName, authenticationType);
             string credentialId = disposableCredential.Credential.Id;
 
             var dataFeedName = Recording.GenerateAlphaNumericId("dataFeed");
@@ -608,11 +617,13 @@ namespace Azure.AI.MetricsAdvisor.Tests
         }
 
         [RecordedTest]
-        [TestCase(SqlServerDataFeedSource.AuthenticationType.Basic)]
-        [TestCase(SqlServerDataFeedSource.AuthenticationType.ManagedIdentity)]
-        public async Task CreateAndGetSqlServerDataFeedWithAuthentication(SqlServerDataFeedSource.AuthenticationType authentication)
+        [TestCase("Basic")]
+        [TestCase("ManagedIdentity")]
+        public async Task CreateAndGetSqlServerDataFeedWithAuthentication(string authenticationType)
         {
             MetricsAdvisorAdministrationClient adminClient = GetMetricsAdvisorAdministrationClient();
+            SqlServerDataFeedSource.AuthenticationType authentication =
+                GetAuthenticationInstance<SqlServerDataFeedSource.AuthenticationType>(authenticationType);
 
             var dataFeedName = Recording.GenerateAlphaNumericId("dataFeed");
             var dataSource = new SqlServerDataFeedSource("mock", "mock")
@@ -630,15 +641,17 @@ namespace Azure.AI.MetricsAdvisor.Tests
         }
 
         [RecordedTest]
-        [TestCase(SqlServerDataFeedSource.AuthenticationType.ServicePrincipal)]
-        [TestCase(SqlServerDataFeedSource.AuthenticationType.ServicePrincipalInKeyVault)]
-        [TestCase(SqlServerDataFeedSource.AuthenticationType.SqlConnectionString)]
-        public async Task CreateAndGetSqlServerDataFeedWithCredentialAuthentication(SqlServerDataFeedSource.AuthenticationType authentication)
+        [TestCase("ServicePrincipal")]
+        [TestCase("ServicePrincipalInKeyVault")]
+        [TestCase("SqlConnectionString")]
+        public async Task CreateAndGetSqlServerDataFeedWithCredentialAuthentication(string authenticationType)
         {
             MetricsAdvisorAdministrationClient adminClient = GetMetricsAdvisorAdministrationClient();
+            SqlServerDataFeedSource.AuthenticationType authentication =
+                GetAuthenticationInstance<SqlServerDataFeedSource.AuthenticationType>(authenticationType);
 
             var credentialName = Recording.GenerateAlphaNumericId("credential");
-            await using var disposableCredential = await DisposableDataSourceCredentialEntity.CreateDataSourceCredentialEntityAsync(adminClient, credentialName, authentication.ToString());
+            await using var disposableCredential = await DisposableDataSourceCredentialEntity.CreateDataSourceCredentialEntityAsync(adminClient, credentialName, authenticationType);
             string credentialId = disposableCredential.Credential.Id;
 
             var dataFeedName = Recording.GenerateAlphaNumericId("dataFeed");
@@ -1235,11 +1248,13 @@ namespace Azure.AI.MetricsAdvisor.Tests
         }
 
         [RecordedTest]
-        [TestCase(AzureBlobDataFeedSource.AuthenticationType.Basic)]
-        [TestCase(AzureBlobDataFeedSource.AuthenticationType.ManagedIdentity)]
-        public async Task UpdateAzureBlobDataFeedWithAuthentication(AzureBlobDataFeedSource.AuthenticationType authentication)
+        [TestCase("Basic")]
+        [TestCase("ManagedIdentity")]
+        public async Task UpdateAzureBlobDataFeedWithAuthentication(string authenticationType)
         {
             MetricsAdvisorAdministrationClient adminClient = GetMetricsAdvisorAdministrationClient();
+            AzureBlobDataFeedSource.AuthenticationType authentication =
+                GetAuthenticationInstance<AzureBlobDataFeedSource.AuthenticationType>(authenticationType);
 
             var dataFeedName = Recording.GenerateAlphaNumericId("dataFeed");
             var dataSource = new AzureBlobDataFeedSource("mock", "mock", "mock");
@@ -1259,11 +1274,13 @@ namespace Azure.AI.MetricsAdvisor.Tests
         }
 
         [RecordedTest]
-        [TestCase(AzureDataExplorerDataFeedSource.AuthenticationType.Basic)]
-        [TestCase(AzureDataExplorerDataFeedSource.AuthenticationType.ManagedIdentity)]
-        public async Task UpdateAzureDataExplorerDataFeedWithAuthentication(AzureDataExplorerDataFeedSource.AuthenticationType authentication)
+        [TestCase("Basic")]
+        [TestCase("ManagedIdentity")]
+        public async Task UpdateAzureDataExplorerDataFeedWithAuthentication(string authenticationType)
         {
             MetricsAdvisorAdministrationClient adminClient = GetMetricsAdvisorAdministrationClient();
+            AzureDataExplorerDataFeedSource.AuthenticationType authentication =
+                GetAuthenticationInstance<AzureDataExplorerDataFeedSource.AuthenticationType>(authenticationType);
 
             var dataFeedName = Recording.GenerateAlphaNumericId("dataFeed");
             var dataSource = new AzureDataExplorerDataFeedSource("mock", "mock");
@@ -1283,14 +1300,16 @@ namespace Azure.AI.MetricsAdvisor.Tests
         }
 
         [RecordedTest]
-        [TestCase(AzureDataExplorerDataFeedSource.AuthenticationType.ServicePrincipal)]
-        [TestCase(AzureDataExplorerDataFeedSource.AuthenticationType.ServicePrincipalInKeyVault)]
-        public async Task UpdateAzureDataExplorerDataFeedWithCredentialAuthentication(AzureDataExplorerDataFeedSource.AuthenticationType authentication)
+        [TestCase("ServicePrincipal")]
+        [TestCase("ServicePrincipalInKeyVault")]
+        public async Task UpdateAzureDataExplorerDataFeedWithCredentialAuthentication(string authenticationType)
         {
             MetricsAdvisorAdministrationClient adminClient = GetMetricsAdvisorAdministrationClient();
+            AzureDataExplorerDataFeedSource.AuthenticationType authentication =
+                GetAuthenticationInstance<AzureDataExplorerDataFeedSource.AuthenticationType>(authenticationType);
 
             var credentialName = Recording.GenerateAlphaNumericId("credential");
-            await using var disposableCredential = await DisposableDataSourceCredentialEntity.CreateDataSourceCredentialEntityAsync(adminClient, credentialName, authentication.ToString());
+            await using var disposableCredential = await DisposableDataSourceCredentialEntity.CreateDataSourceCredentialEntityAsync(adminClient, credentialName, authenticationType);
             string credentialId = disposableCredential.Credential.Id;
 
             var dataFeedName = Recording.GenerateAlphaNumericId("dataFeed");
@@ -1335,15 +1354,17 @@ namespace Azure.AI.MetricsAdvisor.Tests
         }
 
         [RecordedTest]
-        [TestCase(AzureDataLakeStorageDataFeedSource.AuthenticationType.ServicePrincipal)]
-        [TestCase(AzureDataLakeStorageDataFeedSource.AuthenticationType.ServicePrincipalInKeyVault)]
-        [TestCase(AzureDataLakeStorageDataFeedSource.AuthenticationType.SharedKey)]
-        public async Task UpdateAzureDataLakeStorageDataFeedWithCredentialAuthentication(AzureDataLakeStorageDataFeedSource.AuthenticationType authentication)
+        [TestCase("ServicePrincipal")]
+        [TestCase("ServicePrincipalInKeyVault")]
+        [TestCase("SharedKey")]
+        public async Task UpdateAzureDataLakeStorageDataFeedWithCredentialAuthentication(string authenticationType)
         {
             MetricsAdvisorAdministrationClient adminClient = GetMetricsAdvisorAdministrationClient();
+            AzureDataLakeStorageDataFeedSource.AuthenticationType authentication =
+                GetAuthenticationInstance<AzureDataLakeStorageDataFeedSource.AuthenticationType>(authenticationType);
 
             var credentialName = Recording.GenerateAlphaNumericId("credential");
-            await using var disposableCredential = await DisposableDataSourceCredentialEntity.CreateDataSourceCredentialEntityAsync(adminClient, credentialName, authentication.ToString());
+            await using var disposableCredential = await DisposableDataSourceCredentialEntity.CreateDataSourceCredentialEntityAsync(adminClient, credentialName, authenticationType);
             string credentialId = disposableCredential.Credential.Id;
 
             var dataFeedName = Recording.GenerateAlphaNumericId("dataFeed");
@@ -1366,11 +1387,13 @@ namespace Azure.AI.MetricsAdvisor.Tests
         }
 
         [RecordedTest]
-        [TestCase(SqlServerDataFeedSource.AuthenticationType.Basic)]
-        [TestCase(SqlServerDataFeedSource.AuthenticationType.ManagedIdentity)]
-        public async Task UpdateSqlServerDataFeedWithAuthentication(SqlServerDataFeedSource.AuthenticationType authentication)
+        [TestCase("Basic")]
+        [TestCase("ManagedIdentity")]
+        public async Task UpdateSqlServerDataFeedWithAuthentication(string authenticationType)
         {
             MetricsAdvisorAdministrationClient adminClient = GetMetricsAdvisorAdministrationClient();
+            SqlServerDataFeedSource.AuthenticationType authentication =
+                GetAuthenticationInstance<SqlServerDataFeedSource.AuthenticationType>(authenticationType);
 
             var dataFeedName = Recording.GenerateAlphaNumericId("dataFeed");
             var dataSource = new SqlServerDataFeedSource("mock", "mock");
@@ -1390,15 +1413,17 @@ namespace Azure.AI.MetricsAdvisor.Tests
         }
 
         [RecordedTest]
-        [TestCase(SqlServerDataFeedSource.AuthenticationType.ServicePrincipal)]
-        [TestCase(SqlServerDataFeedSource.AuthenticationType.ServicePrincipalInKeyVault)]
-        [TestCase(SqlServerDataFeedSource.AuthenticationType.SqlConnectionString)]
-        public async Task UpdateSqlServerDataFeedWithCredentialAuthentication(SqlServerDataFeedSource.AuthenticationType authentication)
+        [TestCase("ServicePrincipal")]
+        [TestCase("ServicePrincipalInKeyVault")]
+        [TestCase("SqlConnectionString")]
+        public async Task UpdateSqlServerDataFeedWithCredentialAuthentication(string authenticationType)
         {
             MetricsAdvisorAdministrationClient adminClient = GetMetricsAdvisorAdministrationClient();
+            SqlServerDataFeedSource.AuthenticationType authentication =
+                GetAuthenticationInstance<SqlServerDataFeedSource.AuthenticationType>(authenticationType);
 
             var credentialName = Recording.GenerateAlphaNumericId("credential");
-            await using var disposableCredential = await DisposableDataSourceCredentialEntity.CreateDataSourceCredentialEntityAsync(adminClient, credentialName, authentication.ToString());
+            await using var disposableCredential = await DisposableDataSourceCredentialEntity.CreateDataSourceCredentialEntityAsync(adminClient, credentialName, authenticationType);
             string credentialId = disposableCredential.Credential.Id;
 
             var dataFeedName = Recording.GenerateAlphaNumericId("dataFeed");
@@ -2157,6 +2182,14 @@ namespace Azure.AI.MetricsAdvisor.Tests
                     Assert.That(specificDataSource.Query, Is.Null);
                 }
             }
+        }
+
+        private T GetAuthenticationInstance<T>(string authenticationType)
+        {
+            Type dataSourceType = typeof(T);
+            PropertyInfo staticProperty = dataSourceType.GetProperty(authenticationType);
+
+            return (T)staticProperty.GetValue(null);
         }
     }
 }